refactor: move private implementation of g_font_registry to anonymous namespace
